+++ messages.go
@@ -23,193 +23,109 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
-	"strings"
-
 	whttp "github.com/piusalfred/whatsapp/http"
 	"github.com/piusalfred/whatsapp/models"
+	"net/http"
+	"strings"
 )
 
-const ( 
-	// MessageStatusRead is the status for a read message.
-	MessageStatusRead = "read"
+const (
+	MessageEndpoint = "messages"
 )
 
 type (
-	StatusResponse struct {
-		Success bool `json:"success,omitempty"`
-	}
-
-	MessageStatusUpdateRequest struct {
-		MessagingProduct string `json:"messaging_product,omitempty"` // always whatsapp
-		Status           string `json:"status,omitempty"`            // always read
-		MessageID        string `json:"message_id,omitempty"`
+	TextMessage struct {
+		Message    string
+		PreviewURL bool
 	}
 )
 
-type SendTextRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	Message       string
-	PreviewURL    bool
-}
-
-// SendText sends a text message to the recipient.
-func SendText(ctx context.Context, client *http.Client, req *SendTextRequest,
-	hooks ...whttp.Hook,
-) (*ResponseMessage, error) {
+// SendTextMessage sends a text message to a WhatsApp Business Account.
+func (client *Client) SendTextMessage(ctx context.Context, recipient string,
+	message *TextMessage) (*ResponseMessage, error) {
 	text := &models.Message{
 		Product:       messagingProduct,
-		To:            req.Recipient,
+		To:            recipient,
 		RecipientType: individualRecipientType,
 		Type:          textMessageType,
 		Text: &models.Text{
-			PreviewURL: req.PreviewURL,
-			Body:       req.Message,
+			PreviewURL: message.PreviewURL,
+			Body:       message.Message,
 		},
 	}
 
-	reqCtx := &whttp.RequestContext{
-		Name:       "send text",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
-	}
-
-	params := &whttp.Request{
-		Context: reqCtx,
-		Method:  http.MethodPost,
-		Headers: map[string]string{"Content-Type": "application/json"},
-		Query:   nil,
-		Bearer:  req.AccessToken,
-		Form:    nil,
-		Payload: text,
-	}
-
-	var message ResponseMessage
-	err := whttp.Do(ctx, client, params, &message, hooks...)
-	if err != nil {
-		return nil, fmt.Errorf("send text message: %w", err)
-	}
-
-	return &message, nil
+	return client.SendMessage(ctx, "text message", text)
 }
 
-type SendLocationRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	Name          string
-	Address       string
-	Latitude      float64
-	Longitude     float64
-}
+// SendLocationMessage sends a location message to a WhatsApp Business Account.
+func (client *Client) SendLocationMessage(ctx context.Context, recipient string,
+	message *models.Location) (*ResponseMessage, error) {
 
-func SendLocation(ctx context.Context, client *http.Client, req *SendLocationRequest,
-	hooks ...whttp.Hook,
-) (*ResponseMessage, error) {
 	location := &models.Message{
 		Product:       messagingProduct,
-		To:            req.Recipient,
+		To:            recipient,
 		RecipientType: individualRecipientType,
 		Type:          locationMessageType,
 		Location: &models.Location{
-			Name:      req.Name,
-			Address:   req.Address,
-			Latitude:  req.Latitude,
-			Longitude: req.Longitude,
+			Name:      message.Name,
+			Address:   message.Address,
+			Latitude:  message.Latitude,
+			Longitude: message.Longitude,
 		},
-		Contacts:    nil,
-		Interactive: nil,
 	}
 
-	reqCtx := &whttp.RequestContext{
-		Name:       "send location",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
-	}
-
-	params := &whttp.Request{
-		Context: reqCtx,
-		Method:  http.MethodPost,
-		Headers: map[string]string{"Content-Type": "application/json"},
-		Bearer:  req.AccessToken,
-		Payload: location,
-	}
-
-	var message ResponseMessage
-	err := whttp.Do(ctx, client, params, &message, hooks...)
-	if err != nil {
-		return nil, fmt.Errorf("send location: %w", err)
-	}
-
-	return &message, nil
+	return client.SendMessage(ctx, "location message", location)
 }
 
-type ReactRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	MessageID     string
-	Emoji         string
+type ReactMessage struct {
+	MessageID string
+	Emoji     string
 }
 
-/*
-React sends a reaction to a message.
-To send reaction messages, make a POST call to /PHONE_NUMBER_ID/messages and attach a message object
-with type=reaction. Then, add a reaction object.
-
-Sample request:
-
-	curl -X  POST \
-	 'https://graph.facebook.com/v15.0/FROM_PHONE_NUMBER_ID/messages' \
-	 -H 'Authorization: Bearer ACCESS_TOKEN' \
-	 -H 'Content-Type: application/json' \
-	 -d '{
-	  "messaging_product": "whatsapp",
-	  "recipient_type": "individual",
-	  "to": "PHONE_NUMBER",
-	  "type": "reaction",
-	  "reaction": {
-	    "message_id": "wamid.HBgLM...",
-	    "emoji": "\uD83D\uDE00"
-	  }
-	}'
-
-If the message you are reacting to is more than 30 days old, doesn't correspond to any message
-in the conversation, has been deleted, or is itself a reaction message, the reaction message will
-not be delivered, and you will receive a webhooks with the code 131009.
-
-A successful response includes an object with an identifier prefixed with wamid. Use the ID listed
-after wamid to track your message status.
-
-Example response:
-
-	{
-	  "messaging_product": "whatsapp",
-	  "contacts": [{
-	      "input": "PHONE_NUMBER",
-	      "wa_id": "WHATSAPP_ID",
-	    }]
-	  "messages": [{
-	      "id": "wamid.ID",
-	    }]
-	}
-*/
-func React(ctx context.Context, client *http.Client, req *ReactRequest, hooks ...whttp.Hook) (*ResponseMessage, error) {
+// React sends a reaction to a message.
+// To send reaction messages, make a POST call to /PHONE_NUMBER_ID/messages and attach a message object
+// with type=reaction. Then, add a reaction object.
+//
+// Sample request:
+//
+//	curl -X  POST \
+//	 'https://graph.facebook.com/v15.0/FROM_PHONE_NUMBER_ID/messages' \
+//	 -H 'Authorization: Bearer ACCESS_TOKEN' \
+//	 -H 'Content-Type: application/json' \
+//	 -d '{
+//	  "messaging_product": "whatsapp",
+//	  "recipient_type": "individual",
+//	  "to": "PHONE_NUMBER",
+//	  "type": "reaction",
+//	  "reaction": {
+//	    "message_id": "wamid.HBgLM...",
+//	    "emoji": "\uD83D\uDE00"
+//	  }
+//	}'
+//
+// If the message you are reacting to is more than 30 days old, doesn't correspond to any message
+// in the conversation, has been deleted, or is itself a reaction message, the reaction message will
+// not be delivered, and you will receive a webhooks with the code 131009.
+//
+// A successful response includes an object with an identifier prefixed with wamid. Use the ID listed
+// after wamid to track your message status.
+//
+// Example response:
+//
+//	{
+//	  "messaging_product": "whatsapp",
+//	  "contacts": [{
+//	      "input": "PHONE_NUMBER",
+//	      "wa_id": "WHATSAPP_ID",
+//	    }]
+//	  "messages": [{
+//	      "id": "wamid.ID",
+//	    }]
+//	}
+func (client *Client) React(ctx context.Context, recipient string, req *ReactMessage) (*ResponseMessage, error) {
 	reaction := &models.Message{
 		Product: messagingProduct,
-		To:      req.Recipient,
+		To:      recipient,
 		Type:    reactionMessageType,
 		Reaction: &models.Reaction{
 			MessageID: req.MessageID,
@@ -217,93 +133,95 @@ func React(ctx context.Context, client *http.Client, req *ReactRequest, hooks ..
 		},
 	}
 
-	reqCtx := &whttp.RequestContext{
-		Name:       "react",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
-	}
+	return client.SendMessage(ctx, "react to message", reaction)
+}
 
-	params := &whttp.Request{
+func (client *Client) SendMessage(ctx context.Context, name string, msg *models.Message) (*ResponseMessage, error) {
+	cctx := client.Context()
+
+	reqCtx := &whttp.RequestContext{
+		Name:       name,
+		BaseURL:    cctx.BaseURL,
+		ApiVersion: cctx.ApiVersion,
+		SenderID:   cctx.PhoneNumberID,
+		Endpoints:  []string{MessageEndpoint},
+	}
+	request := &whttp.Request{
 		Context: reqCtx,
 		Method:  http.MethodPost,
 		Headers: map[string]string{"Content-Type": "application/json"},
-		Bearer:  req.AccessToken,
-		Payload: reaction,
+		Bearer:  cctx.AccessToken,
+		Payload: msg,
 	}
-
-	var message ResponseMessage
-	err := whttp.Do(ctx, client, params, &message, hooks...)
+	var resp ResponseMessage
+	err := client.http.Do(ctx, request, &resp)
 	if err != nil {
-		return nil, fmt.Errorf("send reaction: %w", err)
+		return nil, fmt.Errorf("%s: %w", name, err)
 	}
-
-	return &message, nil
+	return &resp, nil
 }
 
-type SendContactRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	Contacts      []*models.Contact
-}
-
-func SendContact(ctx context.Context, client *http.Client, req *SendContactRequest,
-	hooks ...whttp.Hook,
-) (*ResponseMessage, error) {
+// SendContacts sends a contact message. Contacts can be easily built using the models.NewContact() function.
+func (client *Client) SendContacts(ctx context.Context, recipient string, contacts []*models.Contact) (
+	*ResponseMessage, error) {
 	contact := &models.Message{
 		Product:       messagingProduct,
-		To:            req.Recipient,
+		To:            recipient,
 		RecipientType: individualRecipientType,
 		Type:          contactsMessageType,
-		Contacts:      req.Contacts,
+		Contacts:      contacts,
 	}
+
+	return client.SendMessage(ctx, "send contacts", contact)
+
+}
+
+// MarkMessageRead sends a read receipt for a message.
+func (client *Client) MarkMessageRead(ctx context.Context, messageID string) (*StatusResponse, error) {
+	reqBody := &MessageStatusUpdateRequest{
+		MessagingProduct: messagingProduct,
+		Status:           MessageStatusRead,
+		MessageID:        messageID,
+	}
+
+	cctx := client.Context()
+
 	reqCtx := &whttp.RequestContext{
-		Name:       "send contacts",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
+		Name:       "mark read",
+		BaseURL:    cctx.BaseURL,
+		ApiVersion: cctx.ApiVersion,
+		SenderID:   cctx.PhoneNumberID,
+		Endpoints:  []string{MessageEndpoint},
 	}
 
 	params := &whttp.Request{
 		Context: reqCtx,
 		Method:  http.MethodPost,
 		Headers: map[string]string{"Content-Type": "application/json"},
-		Query:   nil,
-		Bearer:  req.AccessToken,
-		Form:    nil,
-		Payload: contact,
+		Bearer:  cctx.AccessToken,
+		Payload: reqBody,
 	}
 
-	var message ResponseMessage
-
-	err := whttp.Do(ctx, client, params, &message, hooks...)
+	var success StatusResponse
+	err := client.http.Do(ctx, params, &success)
 	if err != nil {
-		return nil, fmt.Errorf("send contact: %w", err)
+		return nil, fmt.Errorf("client: %w", err)
 	}
 
-	return &message, nil
+	return &success, nil
 }
 
 // ReplyRequest contains options for replying to a message.
 type ReplyRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	Context       string // this is ID of the message to reply to
-	MessageType   MessageType
-	Content       any // this is a Text if MessageType is Text
+	Recipient   string
+	Context     string // this is ID of the message to reply to
+	MessageType MessageType
+	Content     any // this is a Text if MessageType is Text
 }
 
 // Reply is used to reply to a message. It accepts a ReplyRequest and returns a Response and an error.
 // You can send any message as a reply to a previous message in a conversation by including the previous
-// message's ID set as Context in ReplyRequest. The recipient will receive the new message along with a
+// message's ID set as Ctx in ReplyRequest. The recipient will receive the new message along with a
 // contextual bubble that displays the previous message's content.
 //
 // Recipients will not see a contextual bubble if:
@@ -314,7 +232,7 @@ type ReplyRequest struct {
 // Example of Text reply:
 // "messaging_product": "whatsapp",
 //
-//	  "context": {
+//	  "Ctx": {
 //	    "message_id": "MESSAGE_ID"
 //	  },
 //	  "to": "<phone number> or <wa_id>",
@@ -324,7 +242,7 @@ type ReplyRequest struct {
 //	    "body": "your-text-message-content"
 //	  }
 //	}'
-func Reply(ctx context.Context, client *http.Client, request *ReplyRequest,
+func (client *Client) Reply(ctx context.Context, request *ReplyRequest,
 	hooks ...whttp.Hook,
 ) (*ResponseMessage, error) {
 	if request == nil {
@@ -334,11 +252,13 @@ func Reply(ctx context.Context, client *http.Client, request *ReplyRequest,
 	if err != nil {
 		return nil, fmt.Errorf("reply: %w", err)
 	}
+	cctx := client.Context()
+
 	reqCtx := &whttp.RequestContext{
-		Name:       "reply",
-		BaseURL:    request.BaseURL,
-		ApiVersion: request.ApiVersion,
-		SenderID:   request.PhoneNumberID,
+		Name:       "reply to message",
+		BaseURL:    cctx.BaseURL,
+		ApiVersion: cctx.ApiVersion,
+		SenderID:   cctx.PhoneNumberID,
 		Endpoints:  []string{"messages"},
 	}
 
@@ -347,13 +267,13 @@ func Reply(ctx context.Context, client *http.Client, request *ReplyRequest,
 		Method:  http.MethodPost,
 		Headers: map[string]string{"Content-Type": "application/json"},
 		Query:   nil,
-		Bearer:  request.AccessToken,
+		Bearer:  cctx.AccessToken,
 		Form:    nil,
 		Payload: payload,
 	}
 
 	var message ResponseMessage
-	err = whttp.Do(ctx, client, req, &message, hooks...)
+	err = client.http.Do(ctx, req, &message)
 	if err != nil {
 		return nil, fmt.Errorf("reply: %w", err)
 	}
@@ -379,274 +299,6 @@ func formatReplyPayload(options *ReplyRequest) ([]byte, error) {
 	payloadBuilder.WriteString(string(options.MessageType))
 	payloadBuilder.WriteString(`":`)
 	payloadBuilder.Write(contentByte)
-	payloadBuilder.WriteString(`}`)
-
-	return []byte(payloadBuilder.String()), nil
-}
-
-type SendTemplateRequest struct {
-	BaseURL                string
-	AccessToken            string
-	PhoneNumberID          string
-	ApiVersion             string //nolint: revive,stylecheck
-	Recipient              string
-	TemplateLanguageCode   string
-	TemplateLanguagePolicy string
-	TemplateName           string
-	TemplateComponents     []*models.TemplateComponent
-}
-
-func SendTemplate(ctx context.Context, client *http.Client, req *SendTemplateRequest,
-	hooks ...whttp.Hook,
-) (*ResponseMessage, error) {
-	template := &models.Message{
-		Product:       messagingProduct,
-		To:            req.Recipient,
-		RecipientType: individualRecipientType,
-		Type:          templateMessageType,
-		Template: &models.Template{
-			Language: &models.TemplateLanguage{
-				Code:   req.TemplateLanguageCode,
-				Policy: req.TemplateLanguagePolicy,
-			},
-			Name:       req.TemplateName,
-			Components: req.TemplateComponents,
-		},
-	}
-	reqCtx := &whttp.RequestContext{
-		Name:       "send template",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
-	}
-	params := &whttp.Request{
-		Method:  http.MethodPost,
-		Payload: template,
-		Context: reqCtx,
-		Headers: map[string]string{
-			"Content-Type": "application/json",
-		},
-		Bearer: req.AccessToken,
-	}
-	var message ResponseMessage
-	err := whttp.Do(ctx, client, params, &message, hooks...)
-	if err != nil {
-		return nil, fmt.Errorf("send template: %w", err)
-	}
-
-	return &message, nil
-}
-
-/*
-CacheOptions contains the options on how to send a media message. You can specify either the
-ID or the link of the media. Also, it allows you to specify caching options.
-
-The Cloud API supports media http caching. If you are using a link (link) to a media asset on your
-server (as opposed to the ID (id) of an asset you have uploaded to our servers),you can instruct us
-to cache your asset for reuse with future messages by including the headers below
-in your server response when we request the asset. If none of these headers are included, we will
-not cache your asset.
-
-	Cache-Control: <CACHE_CONTROL>
-	Last-Modified: <LAST_MODIFIED>
-	ETag: <ETAG>
-
-# CacheControl
-
-The Cache-Control header tells us how to handle asset caching. We support the following directives:
-
-	max-age=n: Indicates how many seconds (n) to cache the asset. We will reuse the cached asset in subsequent
-	messages until this time is exceeded, after which we will request the asset again, if needed.
-	Example: Cache-Control: max-age=604800.
-
-	no-cache: Indicates the asset can be cached but should be updated if the Last-Modified header value
-	is different from a previous response.Requires the Last-Modified header.
-	Example: Cache-Control: no-cache.
-
-	no-store: Indicates that the asset should not be cached. Example: Cache-Control: no-store.
-
-	private: Indicates that the asset is personalized for the recipient and should not be cached.
-
-# LastModified
-
-Last-Modified Indicates when the asset was last modified. Used with Cache-Control: no-cache. If the Last-Modified value
-is different from a previous response and Cache-Control: no-cache is included in the response,
-we will update our cached apiVersion of the asset with the asset in the response.
-Example: Date: Tue, 22 Feb 2022 22:22:22 GMT.
-
-# ETag
-
-The ETag header is a unique string that identifies a specific apiVersion of an asset.
-Example: ETag: "33a64df5". This header is ignored unless both Cache-Control and Last-Modified headers
-are not included in the response. In this case, we will cache the asset according to our own, internal
-logic (which we do not disclose).
-*/
-type CacheOptions struct {
-	CacheControl string `json:"cache_control,omitempty"`
-	LastModified string `json:"last_modified,omitempty"`
-	ETag         string `json:"etag,omitempty"`
-	Expires      int64  `json:"expires,omitempty"`
-}
-
-type SendMediaRequest struct {
-	BaseURL       string
-	AccessToken   string
-	PhoneNumberID string
-	ApiVersion    string //nolint: revive,stylecheck
-	Recipient     string
-	Type          MediaType
-	MediaID       string
-	MediaLink     string
-	Caption       string
-	Filename      string
-	Provider      string
-	CacheOptions  *CacheOptions
-}
-
-/*
-SendMedia sends a media message to the recipient. To send a media message, make a POST call to the
-/PHONE_NUMBER_ID/messages endpoint with type parameter set to audio, document, image, sticker, or
-video, and the corresponding information for the media type such as its ID or
-link (see MediaInformation http Caching).
-
-Be sure to keep the following in mind:
-  - Uploaded media only lasts thirty days
-  - Generated download URLs only last five minutes
-  - Always save the media ID when you upload a file
-
-Here’s a list of the currently supported media types. Check out Supported MediaInformation Types for more information.
-  - Audio (<16 MB) – ACC, MP4, MPEG, AMR, and OGG formats
-  - Documents (<100 MB) – text, PDF, Office, and OpenOffice formats
-  - Images (<5 MB) – JPEG and PNG formats
-  - Video (<16 MB) – MP4 and 3GP formats
-  - Stickers (<100 KB) – WebP format
-
-Sample request using image with link:
-
-	curl -X  POST \
-	 'https://graph.facebook.com/v15.0/FROM-PHONE-NUMBER-ID/messages' \
-	 -H 'Authorization: Bearer ACCESS_TOKEN' \
-	 -H 'Content-Type: application/json' \
-	 -d '{
-	  "messaging_product": "whatsapp",
-	  "recipient_type": "individual",
-	  "to": "PHONE-NUMBER",
-	  "type": "image",
-	  "image": {
-	    "link" : "https://IMAGE_URL"
-	  }
-	}'
-
-Sample request using media ID:
-
-	curl -X  POST \
-	 'https://graph.facebook.com/v15.0/FROM-PHONE-NUMBER-ID/messages' \
-	 -H 'Authorization: Bearer ACCESS_TOKEN' \
-	 -H 'Content-Type: application/json' \
-	 -d '{
-	  "messaging_product": "whatsapp",
-	  "recipient_type": "individual",
-	  "to": "PHONE-NUMBER",
-	  "type": "image",
-	  "image": {
-	    "id" : "MEDIA-OBJECT-ID"
-	  }
-	}'
-
-A successful response includes an object with an identifier prefixed with wamid. If you are using a link to
-send the media, please check the callback events delivered to your Webhook server whether the media has been
-downloaded successfully.
-
-	{
-	  "messaging_product": "whatsapp",
-	  "contacts": [{
-	      "input": "PHONE_NUMBER",
-	      "wa_id": "WHATSAPP_ID",
-	    }]
-	  "messages": [{
-	      "id": "wamid.ID",
-	    }]
-	}
-*/
-func SendMedia(ctx context.Context, client *http.Client, req *SendMediaRequest,
-	hooks ...whttp.Hook,
-) (*ResponseMessage, error) {
-	if req == nil {
-		return nil, fmt.Errorf("request is nil: %w", ErrBadRequestFormat)
-	}
-
-	payload, err := formatMediaPayload(req)
-	if err != nil {
-		return nil, err
-	}
-
-	reqCtx := &whttp.RequestContext{
-		Name:       "send media",
-		BaseURL:    req.BaseURL,
-		ApiVersion: req.ApiVersion,
-		SenderID:   req.PhoneNumberID,
-		Endpoints:  []string{"messages"},
-	}
-
-	params := &whttp.Request{
-		Context: reqCtx,
-		Method:  http.MethodPost,
-		Bearer:  req.AccessToken,
-		Headers: map[string]string{"Content-Type": "application/json"},
-		Payload: payload,
-	}
-
-	if req.CacheOptions != nil {
-		if req.CacheOptions.CacheControl != "" {
-			params.Headers["Cache-Control"] = req.CacheOptions.CacheControl
-		} else if req.CacheOptions.Expires > 0 {
-			params.Headers["Cache-Control"] = fmt.Sprintf("max-age=%d", req.CacheOptions.Expires)
-		}
-		if req.CacheOptions.LastModified != "" {
-			params.Headers["Last-Modified"] = req.CacheOptions.LastModified
-		}
-		if req.CacheOptions.ETag != "" {
-			params.Headers["ETag"] = req.CacheOptions.ETag
-		}
-	}
-
-	var message ResponseMessage
-
-	err = whttp.Do(ctx, client, params, &message, hooks...)
-	if err != nil {
-		return nil, fmt.Errorf("send media: %w", err)
-	}
-
-	return &message, nil
-}
-
-// formatMediaPayload builds the payload for a media message. It accepts SendMediaOptions
-// and returns a byte array and an error. This function is used internally by SendMedia.
-// if neither ID nor Link is specified, it returns an error.
-func formatMediaPayload(options *SendMediaRequest) ([]byte, error) {
-	media := &models.Media{
-		ID:       options.MediaID,
-		Link:     options.MediaLink,
-		Caption:  options.Caption,
-		Filename: options.Filename,
-		Provider: options.Provider,
-	}
-	mediaJSON, err := json.Marshal(media)
-	if err != nil {
-		return nil, fmt.Errorf("format media payload: %w", err)
-	}
-	recipient := options.Recipient
-	mediaType := string(options.Type)
-	payloadBuilder := strings.Builder{}
-	payloadBuilder.WriteString(`{"messaging_product":"whatsapp","recipient_type":"individual","to":"`)
-	payloadBuilder.WriteString(recipient)
-	payloadBuilder.WriteString(`","type": "`)
-	payloadBuilder.WriteString(mediaType)
-	payloadBuilder.WriteString(`","`)
-	payloadBuilder.WriteString(mediaType)
-	payloadBuilder.WriteString(`":`)
-	payloadBuilder.Write(mediaJSON)
 	payloadBuilder.WriteString(`}`)
 
 	return []byte(payloadBuilder.String()), nil
